+++ prettier.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+    printWidth: 120,
+    semi: false,
+    singleQuote: true,
+    tabWidth: 4,
+    trailingComma: 'es5',
+    overrides: [
+        {
+            files: '*.scss',
+            options: {
+                singleQuote: false,
+            },
+        },
+    ],
+}